requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
pymongo==4.5.0
mysqlclient==2.2.0
psycopg2-binary==2.9.9
